Refactor validators tests (#207)
* Split validators into modules

* Split validators tests. Factor out long fixture.

* Check log messages in tests

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Add bboxes tests and put dataset fixtures in test file

* Fix rebase side effect

* Fix ky -> key

---------
